improved: extract Model::ZipLoader
