+++ src/CMakeLists.txt
@@ -12,6 +12,7 @@ add_executable(runtime
 add_library(vgg_internal OBJECT
   Controller/Controller.cpp
   Exec/VggExec.cpp
+  Model/Loader/ZipLoader.cpp
   Model/JsonDocument.cpp
   Model/JsonSchemaValidator.cpp
   Model/SchemaValidJsonDocument.cpp
