+++ src/Model/Loader/Loader.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <string>
+
 namespace VGG
 {
 namespace Model
@@ -9,7 +11,6 @@ class Loader
 {
 public:
   virtual ~Loader() = default;
-  virtual bool load() = 0;
   virtual bool readFile(const std::string& name, std::string& content) const = 0;
 };
